lots of fixes also uint64 -> float64

+++ tse-p2/globals/globals.go
@@ -1,4 +1,4 @@
 package globals
 
-const TICK_PER_SECOND uint32 = 8 // For speed keep it a power of two.
+const TICK_PER_SECOND uint64 = 8 // For speed keep it a power of two.
